# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TutorMatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bids:
    get:
      operationId: getBids
      tags: [bids]
      summary: List bids
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: query
          schema:
            type: integer
        - name: tutorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, Accepted, Rejected]
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: createBid
      tags: [bids]
      summary: Submit a bid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBidRequest"
      responses:
        "201":
          description: Bid created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
  /bids/{bidId}:
    get:
      operationId: getBid
      tags: [bids]
      summary: Get bid by ID
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
    put:
      operationId: updateBid
      tags: [bids]
      summary: Accept or reject a bid
      security:
        - bearerAuth: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBidRequest"
      responses:
        "200":
          description: Bid updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
...
  /dashboard/student/{studentId}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Get student dashboard summary
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
  /dashboard/tutor/{tutorId}:
    get:
      operationId: getTutorDashboard
      tags: [dashboard]
      summary: Get tutor dashboard summary
      security:
        - bearerAuth: []
      parameters:
        - name: tutorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorDashboard"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [users]
      summary: Admin - list all users
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, tutor, admin]
        - name: suspended
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{userId}/suspend:
    put:
      operationId: adminSuspendUser
      tags: [users]
      summary: Admin - suspend or unsuspend a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserRequest"
      responses:
        "200":
          description: User suspension updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [student, tutor]
      required: [name, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, tutor, admin]
        rating:
          type: number
          nullable: true
        suspended:
          type: boolean
        createdDate:
          type: string
          format: date-time
      required: [userId, name, email, role, suspended, createdDate]
...
    TutorProfile:
      type: object
      properties:
        tutorId:
          type: integer
        subjects:
          type: array
          items:
            type: string
        education:
          type: string
          nullable: true
        experience:
          type: string
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/User"
      required: [tutorId, subjects]
...
    Question:
      type: object
      properties:
        questionId:
          type: integer
        studentId:
          type: integer
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        preferredDuration:
          type: integer
          description: Duration in minutes
        optionalBudget:
          type: number
          nullable: true
        status:
          type: string
          enum: [Open, Matched, Scheduled, Completed, Cancelled]
        createdDate:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
        bidCount:
          type: integer
      required: [questionId, studentId, title, description, subject, preferredDuration, status, createdDate, bidCount]
...
    Bid:
      type: object
      properties:
        bidId:
          type: integer
        questionId:
          type: integer
        tutorId:
          type: integer
        price:
          type: number
        message:
          type: string
        estimatedDuration:
          type: integer
          description: Duration in minutes
        status:
          type: string
          enum: [Pending, Accepted, Rejected]
        createdDate:
          type: string
          format: date-time
        tutor:
          $ref: "#/components/schemas/User"
        tutorProfile:
          $ref: "#/components/schemas/TutorProfile"
      required: [bidId, questionId, tutorId, price, message, estimatedDuration, status, createdDate]
    CreateBidRequest:
      type: object
      properties:
        questionId:
          type: integer
        price:
          type: number
        message:
          type: string
        estimatedDuration:
          type: integer
      required: [questionId, price, message, estimatedDuration]
    UpdateBidRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, Accepted, Rejected]
      required: [status]
    Session:
      type: object
      properties:
        sessionId:
          type: integer
        questionId:
          type: integer
        studentId:
          type: integer
        tutorId:
          type: integer
        proposedTime:
          type: string
          format: date-time
          nullable: true
        tutorCounterTime:
          type: string
          format: date-time
          nullable: true
        finalTime:
          type: string
          format: date-time
          nullable: true
        meetingLink:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending Confirmation, Confirmed, Completed, Cancelled]
        student:
          $ref: "#/components/schemas/User"
        tutor:
          $ref: "#/components/schemas/User"
        question:
          $ref: "#/components/schemas/Question"
      required: [sessionId, questionId, studentId, tutorId, status]
...
    StudentDashboard:
      type: object
      properties:
        openQuestions:
          type: integer
        matchedQuestions:
          type: integer
        scheduledSessions:
          type: integer
        completedSessions:
          type: integer
        totalSpent:
          type: number
        pendingBids:
          type: integer
        recentQuestions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        upcomingSessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
      required: [openQuestions, matchedQuestions, scheduledSessions, completedSessions, totalSpent, pendingBids, recentQuestions, upcomingSessions]
    SuspendUserRequest:
      type: object
      properties:
        suspended:
          type: boolean
      required: [suspended]
    TutorDashboard:
      type: object
      properties:
        openBids:
          type: integer
        acceptedBids:
          type: integer
        scheduledSessions:
          type: integer
        completedSessions:
          type: integer
        totalEarned:
          type: number
        averageRating:
          type: number
        recentBids:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
        upcomingSessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
      required: [openBids, acceptedBids, scheduledSessions, completedSessions, totalEarned, averageRating, recentBids, upcomingSessions]