# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TutorMatch API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from "./question";
import type { Session } from "./session";

export interface StudentDashboard {
  openQuestions: number;
  matchedQuestions: number;
  scheduledSessions: number;
  completedSessions: number;
  totalSpent: number;
  pendingBids: number;
  recentQuestions: Question[];
  upcomingSessions: Session[];
}