# lib/api-zod/src/generated/types/updateBidRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TutorMatch API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBidRequestStatus } from "./updateBidRequestStatus";

export interface UpdateBidRequest {
  status: UpdateBidRequestStatus;
}