# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TutorMatch API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  userId: number;
  name: string;
  email: string;
  role: UserRole;
  rating?: number | null;
  suspended: boolean;
  createdDate: Date;
}